Initial refactor and new component for lock with no key
